refactor: Reducir repositories y APIService
